+++ CHANGELOG.md
@@ -6,6 +6,8 @@ All notable changes to Dagobert are documented here. The format follows
 
 ## [Unreleased]
 
+## [0.5.0] - 2026-09-21
+
 ### Added
 
 - The colour of tags and workflows can now be customised past the builtin palette.